pyproject.toml
[project]
name = "ble-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "crc>=7.1.0",
    "pyserial>=3.5",
    "tqdm>=4.67.3",
]
